Implement CustomerOrder entity and relationship between Client and Station along with CRUD operation for controller and services

+++ src/main/java/com/CarServieStation/backend/service/CarService.java
@@ -1,14 +1,18 @@
 package com.CarServieStation.backend.service;
 
 
+import com.CarServieStation.backend.dto.CarResponseDto;
 import com.CarServieStation.backend.entity.Car;
-import com.CarServieStation.backend.exception.NotFoundOrAlreadyExistException;
+import com.CarServieStation.backend.exception.NotFoundException;
+import com.CarServieStation.backend.exception.OperationNotAllowedException;
 import com.CarServieStation.backend.repository.CarRepository;
+import com.CarServieStation.backend.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CarService {
@@ -16,36 +20,70 @@ public class CarService {
     @Autowired
     private CarRepository carRepository;
 
-    @Transactional
-    public Car createCar(Car car) {
-        return carRepository.save(car);
-    }
+    @Autowired
+    private OrderRepository orderRepository;
 
-
-    public Car getCar(Integer id) {
-        return carRepository.findById(id).orElseThrow(() -> new NotFoundOrAlreadyExistException("Car not found with id: " + id));
-    }
-
-    public List<Car> getAllCars() {
-        return carRepository.findAll();
-    }
 
     @Transactional
-    public Car updateCar(Integer id, Car carDetails) {
-        Car car = getCar(id);
+    public CarResponseDto createCar(Car car) {
+        Car savedCar = carRepository.save(car);
+        return convertToDto(savedCar);
+    }
+
+    public CarResponseDto getCar(Integer id) {
+        Car car = carRepository.findById(id).orElseThrow(() -> new NotFoundException("Car not found with id: " + id));
+        return convertToDto(car);    }
+
+    public List<CarResponseDto> getAllCars() {
+        return carRepository.findAll().stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+    }
+
+    public List<CarResponseDto> getUnassignedCars() {
+        return carRepository.findUnassignedCars().stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public CarResponseDto updateCar(Integer id, Car carDetails) {
+        Car car = getCarEntity(id);
         car.setMake(carDetails.getMake());
         car.setModel(carDetails.getModel());
         car.setColor(carDetails.getColor());
         car.setLicenceNumber(car.getLicenceNumber());
-        return carRepository.save(car);
+
+        Car updatedCar = carRepository.save(car);
+        return convertToDto(updatedCar);
     }
 
 
     @Transactional
     public void deleteCar(Integer id) {
-        Car car = getCar(id);
+        Car car = getCarEntity(id);
+
+        boolean isCarInOrder = orderRepository.existsByLicenceNumber(car.getLicenceNumber());
+        if (isCarInOrder) {
+            throw new OperationNotAllowedException("Car with id " + id + " cannot be deleted as it is associated with an order.");
+        }
+
         car.setClient(null); // Disconnect from client
         carRepository.save(car);
         carRepository.deleteById(id);
+    }
+
+    private Car getCarEntity(Integer id) {
+        return carRepository.findById(id).orElseThrow(() -> new NotFoundException("Car not found with id: " + id));
+    }
+
+    private CarResponseDto convertToDto(Car car) {
+        return new CarResponseDto(
+                car.getId(),
+                car.getModel(),
+                car.getMake(),
+                car.getLicenceNumber(),
+                car.getColor()
+        );
     }
 }
